+++ assets/js/app.js
@@ -1,42 +1,6 @@
 import 'bootstrap/dist/js/bootstrap.bundle';
-import 'superfish/dist/js/superfish'
-import 'jquery-sticky/jquery.sticky'
-import 'jquery.easing/jquery.easing'
-import 'bootstrap-toggle/js/bootstrap-toggle'
 
 $(document).ready(function () {
-
-    $(function () {
-        $('a[href*="#"]:not([href="#"])').click(function () {
-            if (location.pathname.replace(/^\//, '') == this.pathname.replace(/^\//, '') && location.hostname == this.hostname) {
-                var target = $(this.hash);
-                target = target.length ? target : $('[name=' + this.hash.slice(1) + ']');
-
-                if (target.length) {
-                    $('html, body').animate({
-                        scrollTop: target.offset().top
-                    }, 1000, 'easeInOutExpo');
-
-                    if ($(this).parents('.nav-menu').length) {
-                        $('.nav-menu .menu-active').removeClass('menu-active');
-                        $(this).closest('li').addClass('menu-active');
-                    }
-
-                    if ($('body').hasClass('mobile-nav-active')) {
-                        $('body').removeClass('mobile-nav-active');
-                        $('#mobile-nav-toggle i').toggleClass('fa-times fa-bars');
-                        $('#mobile-body-overly').fadeOut();
-                    }
-                    return false;
-                }
-            }
-        });
-    });
-
-    $('.nav-menu').superfish({
-        animation: {opacity: 'show'},
-        speed: 400
-    });
 
     if ($('#nav-menu-container').length) {
         var $mobile_nav = $('#nav-menu-container').clone().prop({id: 'mobile-nav'});
@@ -72,7 +36,6 @@ $(document).ready(function () {
         $("#mobile-nav, #mobile-nav-toggle").hide();
     }
 
-    $("#header").sticky({topSpacing: 0, zIndex: '50'});
 
     $('[data-toggle="tooltip"]').tooltip();
     $('[data-toggle="popover"]').popover();
@@ -80,18 +43,13 @@ $(document).ready(function () {
     $(window).scroll(function () {
         if ($(this).scrollTop() > 100) {
             $('.scrolltop').fadeIn();
-            $('.facebook-messenger').fadeIn();
         } else {
             $('.scrolltop').fadeOut();
-            $('.facebook-messenger').fadeOut();
         }
     });
 
     $('.scrolltop, #logo a').click(function () {
-        $("html, body").animate({
-            scrollTop: 0
-        }, 1000, 'easeInOutExpo');
-        return false;
+        $(window).scrollTop(0);
     });
 
 });